+++ Tasks/UnityBuild/UnityBuildV2/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.5.0 Unreleased]
+## [2.5.1 Unreleased]
 
 ### Added
 
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed `specifyLogFile` option. A log file is now always created and used to stream log to the DevOps console
 - Removed `logFileName` option. The file's name is now always `UnityBuildLog.log` with a timestamp included
 - Removed `editorLogFilePath` output variable. Please use `logsOutputPath` instead
+- Removed custom command lines mode. Please use the new `Unity CMD` task instead
 
 ## [2.4.5]